update & fix

+++ Hi/URsHiPrime.rs
@@ -2,17 +2,17 @@ use std::env;
 use std::time::{Duration, SystemTime};
 
 const MAX_KEEP: usize = 80000;
-static mut prime_array: [u128; MAX_KEEP + 70000] = [2; MAX_KEEP + 70000];
-static mut offset: u128 = 0;
-static mut max_ind: u128 = 0;
-static mut max_prime: u128 = 0;
+static mut prime_array: [usize; MAX_KEEP + 70000] = [2; MAX_KEEP + 70000];
+static mut offset: usize = 0;
+static mut max_ind: usize = 0;
+static mut max_prime: usize = 0;
 
 
 unsafe fn prime_by_euler(page: usize) {
     let mut num: Vec<bool> = vec![false; page];
     for i in 2..page - 1 {
         if !num[i] {
-            max_prime = i as u128;
+            max_prime = i as usize;
             prime_array[(max_ind - offset) as usize] = max_prime;
             max_ind += 1;
         }
@@ -24,14 +24,14 @@ unsafe fn prime_by_euler(page: usize) {
     }
 }
 
-unsafe fn prime_by_eratosthenes(pos :u128, page :usize){
+unsafe fn prime_by_eratosthenes(pos :usize, page :usize){
     let mut num: Vec<bool> = vec![false; page];
     let mut i = 0;
 
-    while prime_array[i] < ((pos + page as u128) as f64).sqrt().ceil() as u128 {
+    while prime_array[i] < ((pos + page as usize) as f64).sqrt().ceil() as usize {
         let p = prime_array[i];
-        let mut j = (pos as f64 /p as f64).ceil() as u128 * p;
-        while j < pos + page as u128 {
+        let mut j = (pos as f64 /p as f64).ceil() as usize * p;
+        while j < pos + page as usize {
             num[j as usize - pos as usize] = true;
             j += p;
         }
@@ -39,19 +39,19 @@ unsafe fn prime_by_eratosthenes(pos :u128, page :usize){
     }
     for i in 0..num.len() {
         if !num[i] {
-            max_prime = pos + i as u128;
+            max_prime = pos + i as usize;
             prime_array[(max_ind - offset) as usize] = max_prime;
             max_ind += 1;
         }
     }
 }
 
-unsafe fn sieve(limit :u128, page :usize){
+unsafe fn sieve(limit :usize, page :usize){
     prime_by_euler(page);
-    for i in 1..limit/page as u128 {
-        prime_by_eratosthenes(page as u128 * i,page);
-        if max_ind > MAX_KEEP as u128 {
-            offset = max_ind - MAX_KEEP as u128;
+    for i in 1..limit/page as usize {
+        prime_by_eratosthenes(page as usize * i,page);
+        if max_ind > MAX_KEEP as usize {
+            offset = max_ind - MAX_KEEP as usize;
         }
     }
 }
@@ -59,7 +59,7 @@ unsafe fn sieve(limit :u128, page :usize){
 fn main() {
     println!("Hi Prime! I'm Rust :-)");
     let args: Vec<String> = env::args().collect();
-    let limit:u128= args[1].parse().unwrap();
+    let limit:usize= args[1].parse().unwrap();
     let page:usize= args[2].parse().unwrap();
     println!("Calculate prime numbers up to {} using partitioned Eratosthenic sieve",limit);
     let start_time = SystemTime::now();
